HW2 done

+++ homeworks/HW2/src/main/java/ru.atom/model/GameObject.java
@@ -7,5 +7,7 @@ public interface GameObject {
     /**
      * Unique id
      */
-    int getId();
+    long getId();
 }
+
+
